update odoo models
